+++ scripts/class/wild-shape-revert.js
@@ -1,1 +1,2 @@
 await HomebrewMacros.revertWildShape(actor, "Wild Shape");
+await HomebrewMacros.revertWildShape(actor, "Elemental Wild Shape");